+++ Awlad_Zamzam.MVC/Services/Interfaces/ICartService.cs
@@ -5,8 +5,8 @@ namespace Awlad_Zamzam.MVC.Services.Interfaces;
 public interface ICartService
 {
     Task<CartViewModel> GetCartAsync();
-    Task AddItemAsync(Guid productId, int quantity, string? note);
-    Task UpdateQuantityAsync(Guid productId, int quantity);
+    Task AddItemAsync(Guid productId, decimal quantity, string? note);
+    Task UpdateQuantityAsync(Guid productId, decimal quantity);
     Task RemoveItemAsync(Guid productId);
     void Clear();
 }
